src/data/usecases/survey-result/save-survey-result/db-save-survey-result.ts
import { SaveSurveyResultRepository, SaveSurveyResult, SurveyResultModel, SaveSurveyResultModel } from './db-save-survey-result-protocols'

export class DbSaveSurveyResult implements SaveSurveyResult {
  constructor (
    private readonly saveSurveyResultRepository: SaveSurveyResultRepository
  ) {}

  async save (saveSurveyResult: SaveSurveyResultModel): Promise<SurveyResultModel> {
    const surveyResult = await this.saveSurveyResultRepository.save(saveSurveyResult)
    return surveyResult
  }
}
